requirements.txt
discord.py==2.7.1
langdetect==1.0.9
python-dotenv==1.2.2
Requests==2.32.5
strip-markdown==1.3